Added agnostic renderer API and support for OpenGL

+++ Orion/src/Orion/Events/MouseEvent.h
@@ -50,7 +50,7 @@ namespace Orion
 	class ORION_API MouseButtonEvent : public Event
 	{
 	public:
-		inline float GetMouseButton() const { return m_Button; }
+		inline int GetMouseButton() const { return m_Button; }
 
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 	protected:
