+++ Orion/src/Orion/Input.h
@@ -12,13 +12,13 @@ namespace Orion
 		inline static bool IsMouseButtonPressed(int keycode) { return s_Instance->IsMouseButtonPressedImpl(keycode); }
 
 		inline static std::pair<float, float> GetMousePosition() { return s_Instance->GetMousePositionImpl(); }
-		inline static int GetMouseX() { return s_Instance->GetMouseXImpl(); }
-		inline static int GetMouseY() { return s_Instance->GetMouseYImpl(); }
+		inline static float GetMouseX() { return s_Instance->GetMouseXImpl(); }
+		inline static float GetMouseY() { return s_Instance->GetMouseYImpl(); }
 
 	protected:
 		virtual std::pair<float, float> GetMousePositionImpl() = 0;
-		virtual int GetMouseXImpl() = 0;
-		virtual int GetMouseYImpl() = 0;
+		virtual float GetMouseXImpl() = 0;
+		virtual float GetMouseYImpl() = 0;
 		virtual bool IsKeyPressedImpl(int keycode) = 0;
 		virtual bool IsMouseButtonPressedImpl(int keycode) = 0;
 	private:
