+++ Orion/src/Platform/Windows/WindowsInput.cpp
@@ -30,16 +30,16 @@ namespace Orion {
 
 		return {xPos,yPos};
 	}
-	int WindowsInput::GetMouseXImpl()
+	float WindowsInput::GetMouseXImpl()
 	{
 		auto [x, y] = GetMousePositionImpl();
 
-		return x;
+		return (float)x;
 	}
-	int WindowsInput::GetMouseYImpl()
+	float WindowsInput::GetMouseYImpl()
 	{
 		auto [x, y] = GetMousePositionImpl();
 
-		return y;
+		return (float)y;
 	}
 }
